+++ index.js
@@ -13,13 +13,29 @@ function walkSync (baseDir, relativePath) {
   var results = []
   var entries = fs.readdirSync(baseDir + '/' + relativePath).sort()
   for (var i = 0; i < entries.length; i++) {
-    var stats = fs.lstatSync(baseDir + '/' + relativePath + entries[i])
-    if (stats.isDirectory()) {
-      results.push(relativePath + entries[i] + '/')
-      results = results.concat(walkSync(baseDir, relativePath + entries[i]))
+    var entryRelativePath = relativePath + entries[i]
+    var stats = getStat(baseDir + '/' + entryRelativePath)
+
+    if (stats && stats.isDirectory()) {
+      results.push(entryRelativePath + '/')
+      results = results.concat(walkSync(baseDir, entryRelativePath))
     } else {
-      results.push(relativePath + entries[i])
+      results.push(entryRelativePath)
     }
   }
   return results
+}
+
+function getStat(path) {
+  var stat
+
+  try {
+    stat = fs.statSync(path)
+  } catch(error) {
+    if (error.code !== 'ENOENT') {
+      throw error
+    }
+  }
+
+  return stat;
 }